% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{aggregate_knn}
\alias{aggregate_knn}
\title{Aggregate point-wise distances over k nearest neighbors}
\usage{
aggregate_knn(data, fun, k)
}
\arguments{
\item{data}{A numeric matrix, data frame or distance matrix}

\item{fun}{A function to aggregate point-wise distances}

\item{k}{An integer specifying number of nearest neighbors used to compute aggregation}
}
\value{
A numeric vector
}
\description{
Aggregate point-wise distances over k nearest neighbors
}
\examples{
aggregate_knn(iris[, -5], mean, 3) # average 3-nn
aggregate_knn(dist(iris[, -5]), max, 5) # equivalent to exact 5-nn
}
