% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_mvn.R
\name{chisq_mvn}
\alias{chisq_mvn}
\title{Compute multivariate Mahalanobis distances and corresponding chi-squared quantiles
to assess multivariate normality}
\usage{
chisq_mvn(data)
}
\arguments{
\item{data}{a numeric matrix or data.frame}
}
\value{
a Nx2 tibble with observed Mahalanobis distances and ordered chi-sq
quantiles
}
\description{
Compute multivariate Mahalanobis distances and corresponding chi-squared quantiles
to assess multivariate normality
}
\details{
This function is inspired by a discussion of multivariate normality in
"An R and S-Plus Companion to Multivariate Analysis" (Everitt, 2005)
}
\examples{
mvn_coords <- chisq_mvn(iris[, -5])
plot(distances ~ chisq, data = mvn_coords)
abline(lm(distances ~ chisq, data = mvn_coords))
}
