% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{chisq_plot}
\alias{chisq_plot}
\title{Chi-Squared MVN Plot}
\usage{
chisq_plot(data, plot_line = FALSE, line_col = "black", ...)
}
\arguments{
\item{data}{a matrix or data.frame.}
}
\description{
Chi-Squared MVN Plot
}
