% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_centers.R
\name{dist_to_centers}
\alias{dist_to_centers}
\title{Distance to centroid}
\usage{
dist_to_centers(data, labels, centers)
}
\arguments{
\item{data}{data}

\item{labels}{labels}

\item{centers}{centers}
}
\value{
a matrix
}
\description{
Distance to centroid
}
\examples{
x
}
\references{
section 3.2.1 of "Outlier Analysis" (C. C. Aggarwal. Outlier Analyis. 
Springer, 2017.)
}
