% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvr_ensemble.R
\name{dvr_ensemble}
\alias{dvr_ensemble}
\title{Dependent Variable Regression Ensemble}
\usage{
dvr_ensemble(formula, data, method = "lm", n_predictions, n_train_points,
  score_set = c("all", "test"), error_agg_fun = mean, scores_only = TRUE,
  ...)
}
\arguments{
\item{formula}{a formula interface specifying the model (see help("lm") for more detail)}

\item{data}{a matrix or data.frame containing variables in model}

\item{method}{a model function (e.g. "lm", "randomForest")}

\item{n_predictions}{an integer specifying the number of components in the ensemble.
If score_set = "test", set this high enough to ensure all points are predicted
a sufficient number of times}

\item{n_train_points}{an integer or numeric value specifying the number of rows
used in the training phase of each ensemble}

\item{score_set}{one of "all" or "test". If "all", scores all N points in each
training iteration. If "test", score out of sample points in each iteration.}

\item{error_agg_fun}{a function for combining the squared prediction errors.
Defaults to mean.}

\item{scores_only}{logical, if TRUE return a vector of outlier scores. If FALSE,
return the error matrix and outlier scores}
}
\description{
Dependent Variable Regression Ensemble
}
\examples{
ensemble_lm(formula = Sepal.Length ~ ., data = iris[,-5],
n_predictions=100, n_train_points=100, error_agg_fun = median, 
scores_only = T)#' 
}
\references{
section 3.2.1 of "Outlier Analysis" (C. C. Aggarwal. Outlier Analyis. 
Springer, 2017.)
}
