% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{exact_knn}
\alias{exact_knn}
\title{Compute exact k nearest neighbor distance}
\usage{
exact_knn(data, k = 1)
}
\arguments{
\item{data}{A numeric matrix, data frame, or distance matrix}

\item{k}{An integer specifying the k nearest neighbor}
}
\value{
A numeric vector
}
\description{
Compute exact k nearest neighbor distance
}
\examples{
exact_knn(iris[, -5], 3) # 3-nn
}
