% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{Return point-wise k nearest neighbor}
\usage{
nearest_neighbors(distance_matrix, k)
}
\arguments{
\item{distance_matrix}{A distance matrix}

\item{k}{An integer or sequence of integers specifying the desired nearest neighbors}
}
\value{
An N x k tibble with ordered pairwise distances
}
\description{
Return point-wise k nearest neighbor
}
\details{
A distance matrix (either a full square matrix or a "dist" object) is
required as input.
}
\examples{
my_dist <- dist(scale(state.x77))
nearest_neighbors(distance_matrix = my_dist, k = 1:3)
}
