% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{one_hot}
\alias{one_hot}
\title{Categorical Data Matrix to One-Hot Binary Matrix}
\usage{
one_hot(data)
}
\arguments{
\item{data}{a matrix or data.frame of categorical variables as character 
strings or factors.}

\item{normalize}{logical, if FALSE then binary matrix is returned. If TRUE, then
normalization (see details) is applied to each binary transformed variable.}
}
\value{
A transformed matrix is returned.
}
\description{
Categorical Data Matrix to One-Hot Binary Matrix
}
\details{
The normalization technique is taken from Outlier Analysis (Aggarwal, 2017),
section 8.3. For each column j in the binary transformed matrix, a normalization 
factor is defined as sqrt(ni \* pj \* (1-pj)), where ni is the number of distinct
categories in the reference variable from the raw data set and pj is the proportion
of records taking the value of 1 for the jth variable
}
\examples{
x <- data.frame(gender = sample(c("male", "female"), 15, T),
                age_cat = sample(c("young", "old", "unknown"), 15, T)) 
one_hot(data = x, normalize = TRUE)
}
