% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotated_bagging.R
\name{pca_bag}
\alias{pca_bag}
\title{PCA Rotated Bagging}
\usage{
pca_bag(data, n_features = floor(sqrt(ncol(data))), n_iterations = 50,
  outlier_fun, scores_fun = sum, scores_only = TRUE)
}
\arguments{
\item{data}{a matrix or data.frame.}
}
\description{
PCA Rotated Bagging
}
