% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_lm.R
\name{regression_ensemble}
\alias{regression_ensemble}
\title{Regression Ensemble}
\usage{
regression_ensemble(formula, data, method = "lm", n_components,
  n_train_points, error_agg_fun, scores_only = TRUE, ...)
}
\arguments{
\item{formula}{a formula interface specifying the model (see help("lm") for more detail)}

\item{data}{a matrix or data.frame containing variables in model}

\item{method}{a model function (e.g. "lm", "randomForest"). Functions in packages 
outside of the base R distribution are accessed by loading the necessary package}

\item{n_components}{an integer specifying the number of components in the ensemble}

\item{n_train_points}{an integer or numeric value specifying the number of rows
used in the training phase of each ensemble}

\item{error_agg_fun}{a function for combining the squared prediction errors 
(.e.g. "mean", mean, median)}

\item{scores_only}{logical, if TRUE return a vector of outlier scores. If FALSE,
return the error matrix and outlier scores}
}
\description{
Regression Ensemble
}
\examples{
ensemble_lm(formula = Sepal.Length ~ ., data = iris[,-5],
n_components=100, n_train_points=100, error_agg_fun = median, 
scores_only = T)#' 
}
