% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{aggregate_knn}
\alias{aggregate_knn}
\title{Exact K-NN}
\usage{
aggregate_knn(data, fun, k)
}
\arguments{
\item{fun}{A function to aggregate distances}

\item{k}{An integer specifying value for K}

\item{distance_matrix}{A matrix or data frame or distance matrix}
}
\value{
A numeric vector
}
\description{
Exact K-NN
}
\examples{
aggregate_knn(iris[, -5], mean, 3) # average 3-NN
aggregate_knn(dist(iris[, -5]), median, 5)
}
