% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/also.R
\name{also}
\alias{also}
\title{Attribute-wise Learning for Scoring Outliers.}
\usage{
also(data, scale_numerics = TRUE, method, cross_validate = TRUE,
  n_folds = 5, scores_only = TRUE, ...)
}
\arguments{
\item{data}{a matrix or data.frame.}

\item{scale_numerics}{logical, if TRUE then center and scale numeric variables.
If FALSE then ignore scaling}

\item{method}{a function name in the form of a character string which is 
passed to formula (e.g. 'lm')}

\item{cross_validate}{logical indicating the use of cross validation for scoring}

\item{n_folds}{an integer specifying the number of folds if cross validating.
Defaults to 5}

\item{scores_only}{logical, if TRUE return outlier scores only. If FALSE
return a list with outlier scores and the error matrix}

\item{...}{additional arguments passed to method formula}
}
\value{
If scores_only = TRUE, only outlier scores are returned. If FALSE, 
the function returns a list containing outlier scores, the error matrix, 
the rmse_matrix, and feature weights.
}
\description{
Attribute-wise Learning for Scoring Outliers.
}
\examples{
# Build ncol(data) linear regression models. In each iteration, a new feature
# is treated as the target and the remaining features are the predictors. With 
# cross_validate == TRUE, 10-fold cross validation is applied in each iteration. All points are 
# therefore given an out of sample score.

also(data = scale(state.x77), method = 'lm', cross_validate = TRUE, n_folds = 10, 
     return_list = TRUE)

# Outlier scores from ncol(data) random forests
 
also(data = scale(state.x77), method = 'randomForest', cross_validate = FALSE, 
     return_list = FALSE)
}
\references{
see "Outlier Analysis" (C.C Aggarwal. Springer, 2017)
}
