% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/also.R
\name{also}
\alias{also}
\title{Attribute-wise Learning for Scoring Outliers.}
\usage{
also(data, method, cv = FALSE, folds = NULL, return_list = TRUE, ...)
}
\arguments{
\item{data}{a matrix or data.frame.}

\item{method}{a function name in the form of a character string which is 
passed to formula (e.g. 'lm')}

\item{cv}{logical indicating the use of cross validation for scoring}

\item{folds}{an integer specifying the number of folds if cv = TRUE}

\item{return_list}{logical, if TRUE return list of output (see details). If 
FALSE return outlier scores only}

\item{...}{additional arguments passed to method formula}
}
\value{
If return_list = TRUE, the function returns a list containing outlier
scores, the error matrix, the rmse_matrix, and feature weights. If return_list
= FALSE, outlier scores only are returned.
}
\description{
Attribute-wise Learning for Scoring Outliers.
}
\examples{
# Build ncol(data) linear regression models. In each iteration, a new feature
# is treated as the target and the remaining features are the predictors. With 
# cv == TRUE, 10-fold cross validation is applied in each iteration. All points are 
# therefore given an out of sample score.

also(data = scale(state.x77), method = 'lm', cv = TRUE, folds = 10, 
     return_list = TRUE)

# Outlier scores from ncol(data) random forests
 
also(data = scale(state.x77), method = 'randomForest', cv = FALSE, 
     return_list = FALSE)
}
