% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{chisq_mvn}
\alias{chisq_mvn}
\title{Chi-Squared Coordinates for Multivariate Normality}
\usage{
chisq_mvn(data)
}
\arguments{
\item{data}{a matrix or data.frame.}
}
\description{
Chi-Squared Coordinates for Multivariate Normality
}
