% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{exact_knn}
\alias{exact_knn}
\title{Exact K-NN}
\usage{
exact_knn(data, k = 1)
}
\arguments{
\item{k}{An integer specifying value for K}

\item{distance_matrix}{A matrix or data frame or distance matrix}
}
\value{
A numeric vector
}
\description{
Exact K-NN
}
\examples{
exact_knn(iris[, -5], 3) # 3-NN
}
