% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{Unsupervised K-Nearest Neighbors}
\usage{
nearest_neighbors(distance_matrix, k)
}
\arguments{
\item{distance_matrix}{A distance matrix}

\item{k}{Value for k. Either an integer or sequence of integers}
}
\value{
A data.frame
}
\description{
Unsupervised K-Nearest Neighbors
}
\details{
A distance matrix (either a full square matrix or a "dist" object) is
required as input.
}
\examples{
my_dist <- dist(scale(state.x77))
nearest_neighbors(distance_matrix = my_dist, k = 1:3)
}
